requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
mysqlclient==2.1.1
sqlalchemy==2.0.7
python-dotenv==1.0.0
pydantic==1.10.7
email-validator==1.3.1
bcrypt==4.0.1 
passlib==1.7.4
cryptography==40.0.1
python-jose==3.3.0
python-multipart==0.0.6